Set tick positions and horizontal alignment
This patch allows to define on which axes ticks are displayed
and to define the horizontal alignment of the labels. In addition, it
is possible to pass the width of the labels to the save function as
this value is required for pgfplots to work correctly.
The value can be passed as the 'x tick label text width'
and 'y tick label text width' in the extra dictionary.

In addition new unit tests are provided.