List all artists in legend
Previously, only line plots were listed in the legend because other
handles to other artists (listed as rectangle by mpl) don't have a
get_color() method and hence raise an AttributeError which is silently
caught.
Now, it is checked for the type of the handle and the appropriate color
getter method is called.