Support annotation.textcoords == "offset points"
This modifies `_annotation` to return the location at which the text should be rendered, and then adds support for `offset points` and `offset pixels`.

This should make it easier to add support for the other coordinate type too, but I'll leave that for someone who needs them.

This also fixes a bug where an arrow would be drawn even if none was requested.
A test is added for this.